added major changes regarding dynamic SEO
